+++ spade-ast-lowering/src/lib.rs
@@ -2561,7 +2561,7 @@ pub fn visit_expression(e: &ast::Expression, ctx: &mut Context) -> Result<hir::E
         ast::Expression::ArrayShorthandLiteral(expr, amount) => {
             Ok(hir::ExprKind::ArrayShorthandLiteral(
                 Box::new(visit_expression(expr, ctx)?.at_loc(expr)),
-                amount.clone(),
+                visit_const_generic(amount, ctx)?.map(|c| c.with_id(ctx.idtracker.next())),
             ))
         }
         ast::Expression::Index(target, index) => {